+++ unpacked/_next/static/[buildHash]/_buildManifest.js
@@ -96,7 +96,7 @@
   "static/chunks/68a27ff6-b1db347c50639918.js",
   "static/chunks/984-1278472924e49180.js",
   "static/chunks/653-4cb05ee80f77e17d.js",
-  "static/chunks/496-7e3343e030973b2c.js",
+  "static/chunks/496-e25244fd42246575.js",
   "static/chunks/2802bd5f-15923fb46be55b45.js",
   "static/chunks/bd26816a-7ae54dd3357d90b4.js",
   "static/chunks/597-ed3dbaa450c308e5.js",